Add files via upload
An analysis of groundwater well periodicity in 6 wells in the Santa Ana Subbasin of California. This is an important factor in determining sustainable yield, which determines the total amount of tradable water in a given basin. Periodicity gives an account of the longer term cycles in groundwater levels, reflective of drought period of approximately 4-6 years or more.